LLM_Detect_master/LLM_Detect_master/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.5
blinker==1.9.0
captcha==0.7.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
Deprecated==1.3.1
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Limiter==4.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.4
openai==2.7.2
openpyxl==3.1.5
ordered-set==4.1.0
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pillow==12.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.1


